Fix refactoring
