+++ neuro_extras/__init__.py
@@ -1,7 +1,5 @@
 from neuromation.api import PluginManager
 
-from .main import main  # noqa
-
 
 # NOTE: When updating the version, don't forget to update main.NEURO_EXTRAS_IMAGE_TAG
 __version__ = "v20.9.30.2"
