+++ neuro_extras/cli.py
@@ -1,0 +1,26 @@
+import logging
+
+import click
+
+
+class ClickLogHandler(logging.Handler):
+    def emit(self, record: logging.LogRecord) -> None:
+        try:
+            msg = self.format(record)
+            click.echo(msg)
+        except Exception:
+            self.handleError(record)
+
+
+@click.group()
+def main() -> None:
+    handler = ClickLogHandler()
+    handler.setFormatter(logging.Formatter("%(levelname)s: %(message)s"))
+
+    root_logger = logging.getLogger()
+    root_logger.setLevel(logging.INFO)
+    root_logger.addHandler(handler)
+
+
+NEURO_EXTRAS_IMAGE_TAG = "v20.9.30a7"
+NEURO_EXTRAS_IMAGE = f"neuromation/neuro-extras:{NEURO_EXTRAS_IMAGE_TAG}"
